fix(layering): move graph context and edges to src/runtime (#1948, #2049 Slice A)
Move GraphContext (telegram_bot/graph/context.py) and edge routing
functions (telegram_bot/graph/edges.py) to src/runtime/graph/ as the
first reverse-layering cleanup slice.

Changes:
- src/runtime/graph/context.py: canonical home for GraphContext TypedDict
- src/runtime/graph/edges.py: canonical home for 5 edge routing functions,
  now importing observe from src.observability instead of telegram_bot
- telegram_bot/graph/context.py: back-compat re-export shim
- telegram_bot/graph/edges.py: back-compat re-export shim
- telegram_bot/graph/graph.py: imports GraphContext and edges from
  src.runtime.graph.*
- 6 node files: import GraphContext from src.runtime.graph.context

No runtime behaviour change. All existing imports through
telegram_bot.graph.{context,edges} continue to work through the shims.
The known_layering_violations.json allowlist is unchanged (the remaining
src/api/main.py -> telegram_bot.graph.graph violation requires follow-up
slices that migrate node implementations).

486 graph tests pass. Layering contract test passes.
