+++ telegram_bot/graph/nodes/classify.py
@@ -15,7 +15,7 @@ from typing import Any
 
 from langgraph.runtime import Runtime
 
-from telegram_bot.graph.context import GraphContext
+from src.runtime.graph.context import GraphContext
 from telegram_bot.observability import get_client, observe
