+++ telegram_bot/graph/nodes/guard.py
@@ -17,7 +17,7 @@ from typing import Any
 
 from langgraph.runtime import Runtime
 
-from telegram_bot.graph.context import GraphContext
+from src.runtime.graph.context import GraphContext
 from telegram_bot.observability import get_client, observe
