+++ telegram_bot/graph/nodes/rerank.py
@@ -12,7 +12,7 @@ from typing import Any
 
 from langgraph.runtime import Runtime
 
-from telegram_bot.graph.context import GraphContext
+from src.runtime.graph.context import GraphContext
 from telegram_bot.observability import get_client, observe
 from telegram_bot.services.metrics import PipelineMetrics
 from telegram_bot.services.rag_core import perform_rerank
