+++ telegram_bot/graph/nodes/retrieve.py
@@ -15,7 +15,7 @@ from typing import Any
 
 from langgraph.runtime import Runtime
 
-from telegram_bot.graph.context import GraphContext
+from src.runtime.graph.context import GraphContext
 from telegram_bot.observability import get_client, observe
 from telegram_bot.services.coverage_mode import cap_results_per_doc, detect_coverage_mode
 from telegram_bot.services.metrics import PipelineMetrics
